backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
python-dotenv==1.1.0
django-cors-headers==4.7.0
psycopg2-binary==2.9.10